move `AccountField` into its own file
